# api/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "reverie",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/timeline": {
      "get": {
        "summary": "オープンのタイムラインを取得する",
        "deprecated": false,
        "description": "",
        "operationId": "getTimeline",
        "tags": [],
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "description": "次のページを取得するためのキー",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "description": "正常に処理を終了したかどうか"
                    },
                    "code": {
                      "type": "integer",
                      "description": "レスポンスコード"
                    },
                    "data": {
                      "description": "データ",
                      "$ref": "#/components/schemas/PostTimeline"
                    }
                  },
                  "required": [
                    "ok",
                    "code",
                    "data"
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/timeline/users/{user_id}": {
      "get": {
        "summary": "ユーザーのタイムラインを取得する",
        "deprecated": false,
        "description": "",
        "operationId": "getUserTimeline",
        "tags": [],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "次のページを取得するためのキー",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "description": "正常に処理を終了したかどうか"
                    },
                    "code": {
                      "type": "integer",
                      "description": "レスポンスコード"
                    },
                    "data": {
                      "description": "データ",
                      "$ref": "#/components/schemas/PostTimeline"
                    }
                  },
                  "required": [
                    "ok",
                    "code",
                    "data"
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/timeline/following": {
      "get": {
        "summary": "フォロー中のタイムラインを取得する",
        "deprecated": false,
        "description": "",
        "operationId": "getFollowingTimeline",
        "tags": [],
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "description": "次のページを取得するためのキー",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "description": "正常に処理を終了したかどうか"
                    },
                    "code": {
                      "type": "integer",
                      "description": "レスポンスコード"
                    },
                    "data": {
                      "description": "データ",
                      "$ref": "#/components/schemas/PostTimeline"
                    }
                  },
                  "required": [
                    "ok",
                    "code",
                    "data"
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID番号",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "description": "名前"
          },
          "nickname": {
            "type": "string"
          },
          "avatar_image_url": {
            "type": "string",
            "nullable": true
          },
          "banner_image_url": {
            "type": "string",
            "nullable": true
          },
          "biography": {
            "type": "string",
            "nullable": true
          },
          "is_private": {
            "type": "boolean"
          },
          "block_status": {
            "$ref": "#/components/schemas/BlockStatus"
          },
          "social_connection": {
            "$ref": "#/components/schemas/SocialConnection"
          },
          "social_engagement": {
            "$ref": "#/components/schemas/SocialEngagement"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "name",
          "nickname",
          "avatar_image_url",
          "banner_image_url",
          "biography",
          "updated_at",
          "created_at",
          "is_private"
        ]
      },
      "SocialConnection": {
        "type": "object",
        "properties": {
          "following": {
            "type": "boolean"
          },
          "followed_by": {
            "type": "boolean"
          }
        },
        "required": [
          "following",
          "followed_by"
        ]
      },
...
      "BlockStatus": {
        "type": "object",
        "properties": {
          "blocking": {
            "type": "boolean"
          },
          "blocked_by": {
            "type": "boolean"
          }
        },
        "required": [
          "blocking",
          "blocked_by"
        ]
      },
      "SocialEngagement": {
        "type": "object",
        "properties": {
          "following_count": {
            "type": "integer"
          },
          "followers_count": {
            "type": "integer"
          },
          "posts_count": {
            "type": "integer"
          },
          "media_count": {
            "type": "integer"
          },
          "favorites_count": {
            "type": "integer"
          }
        },
        "required": [
          "following_count",
          "followers_count",
          "posts_count",
          "media_count",
          "favorites_count"
        ]
      },
...
      "Post": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID番号",
            "format": "uuid"
          },
          "author": {
            "$ref": "#/components/schemas/User"
          },
          "text": {
            "type": "string",
            "nullable": true
          },
          "favorited": {
            "type": "boolean"
          },
          "favorites_count": {
            "type": "integer"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "author",
          "updated_at",
          "created_at",
          "text",
          "favorites_count",
          "favorited"
        ]
      },
...
      "PostTimeline": {
        "type": "object",
        "properties": {
          "posts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Post"
            }
          },
          "next_cursor": {
            "type": "string",
            "description": "次のページを取得するためのキー",
            "format": "uuid"
          }
        },
        "required": [
          "posts",
          "next_cursor"
        ]
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "servers": []
}